requirements.txt

flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
openai==1.3.8
anthropic==0.7.8
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
schedule==1.2.0
discord-webhook==1.3.0
yfinance==0.2.28
alpha-vantage==2.3.1
finnhub-python==2.4.19
twelvedata==1.2.11
